+++ app/Http/Shop/Controllers/PhoneController.php
@@ -9,6 +9,7 @@ use App\Http\Shop\Models\Phone;
 use App\Http\Shop\Models\Shop;
 use App\Support\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Log;
 
 class PhoneController extends Controller
@@ -21,16 +22,19 @@ class PhoneController extends Controller
 
     public function index(Shop $shop)
     {
+
+        if ($shop->id !== Auth::guard('shop')->id())
+            return JsonResponse::sendJsonResponse(0, 'خطا', 'کد وارد شده متعلق به فروشگاه شما نیست !!',);
+
         $phones = Phone::where('shop_id', $shop->id)->get();
         return response()->json(['phones' => $phones], 200);
     }
 
     public function store(Request $request, Shop $shop)
     {
-//        dd($request->all());
-//        if (!Auth::user()->can('read-phone')) {
-//            return back()->with('error-message', 'دسترسی شما به این بخش محدود می باشد!');
-//        }
+        if ($shop->id !== Auth::guard('shop')->id())
+            return JsonResponse::sendJsonResponse(0, 'خطا', 'کد وارد شده متعلق به فروشگاه شما نیست !!',);
+
         $request->validate([
             'number' => ['required','digits_between:4,11'],// 'unique:phones,number,' . auth()->id()
         ]);
@@ -42,17 +46,14 @@ class PhoneController extends Controller
 
         return response()->json(['data' => ['success' => true, 'message' => 'با موفقیت ثبت شد','phone_id'=>$phone->id,'phone_number'=>$phone->number]]);
 
-//        return JsonResponse::sendJsonResponse(1, 'موفق', 'شماره تلفن با موفقیت ثبت گردید',
-//            'DATATABLE_REFRESH');
 
     }
 
 
     public function update(Request $request, Phone $phone)
     {
-//        if (!Auth::user()->can('update-phone')) {
-//            return back()->with('error-message', 'دسترسی شما به این بخش محدود می باشد!');
-//        }
+        if ($phone->shop_id !== Auth::guard('shop')->id())
+            return JsonResponse::sendJsonResponse(0, 'خطا', 'کد وارد شده متعلق به فروشگاه شما نیست !!',);
 
         $request->validate([
             'number' => ['required','digits_between:6,11'],// 'unique:phones,number,' . auth()->id()
@@ -65,9 +66,8 @@ class PhoneController extends Controller
 
     public function destroy(Phone $phone)
     {
-//        if (!Auth::user()->can('delete-phone')) {
-//            return back()->with('error-message', 'دسترسی شما به این بخش محدود می باشد!');
-//        }
+        if ($phone->shop_id !== Auth::guard('shop')->id())
+            return JsonResponse::sendJsonResponse(0, 'خطا', 'کد وارد شده متعلق به فروشگاه شما نیست !!',);
 
         $phone->delete();
         return response()->json(['data' => ['success' => true, 'message' => 'با موفقیت حذف شد']]);
